Reworked IgProperty to reduce amount of code to write.
At the moment the only way to construct an IgProperty
is to use the IgCollection::getProperty(name) method.
This patch introduces a new constructor for IgProperty

IgProperty(IgCollection *collection, const char *name)

which internally does the same thing, but reduces the 
amount of code to be written from

IgProperty POS = collection->getProperty("position");

to 

IgProperty POS(collection, "position");

This allows to collapse a few lines and simplify the 
code.