Refactoring of the way associations in ig files are created / iterated.
Various iterator classes were renamed so that they now at least resembles STL iterators. I.e. IgCollection::iterator in place of IgCollectionIterator.

There is now a new iteration syntax that removes the need for an IgAssociatedSet by exploiting the fact that associations are always created in order.

This means that rather than iterating with:

    for (IgCollection::iterator ci = coll.begin(), ce = coll.end();
         ci != ce;
         ++ci)
    {
      IgAssociatedSet &view = assoc.getAssociatedSet(ci)
      for (IgAssociatedSet::iterator ai = view.begin(), ae = view.end(); 
           ai != ae;
           ++ai)
      {..};
    }

the syntax to iterate over associated objects is:

    for (IgCollection::iterator ci = coll.begin(), ce = coll.end();
         ci != ce;
         ++ci)
    {
      for (IgAssociations::iterator ai = assoc.begin(ci), ae = assoc.end(); 
           ai != ae;
           ++ai)
      {..};
    }

Notice that IgAssociations::associate method creates mono-directional and not bi-directional associations. If you want to have the same behavior as before (which was not actually ever used), you need to do:

    associate(a, b); // associates b to a
    associate(b, a); // associates a to b

All the tests have been updated for the new API and they now use QtTest framework.

All the code in `ISpyApplication` was updated to match the changes.

Given that there is no ambiguity anymore, IgAssociationSet was renamed IgAssociations.