Remove hackish code.
I never liked the way the following features where implemented:

* Filtering.
* Auto-play.
* Auto-printing.

In particular I don't like the fact that they relied on 3/4 different timers
and that filtering was part of the event loop, rather than a index to lookup
to decide which event was next.

I simply went ahead and cleaned up the code.